move store into dedicated folder
