+++ src/views/profile/ProfilePage.vue
@@ -11,7 +11,7 @@
 
 <script lang="ts">
 import {Component, Vue} from 'vue-property-decorator';
-import {LOAD_CURRENT_PROFILE} from '@/store.user-profile';
+import {LOAD_CURRENT_PROFILE} from '@/store/store.user-profile';
 import Box from '@/components/Box.vue';
 import TheContact from '@/components/TheContact.vue';
 import {UserProfile} from '@/types';
